// packages/event-handlers/src/types/generated/trainingRequest.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export type TrainingId = string
export type TrainingUnits = number
export type ParentMemoryNodeConfig = RootNode | TrainedNode
export type MemoryId = string
export type MemoryUrl = string
export type Size = number
export type Strength = number
export type MaxSpeed = number
export type Endurance = number
/**
 * Configuration for an opponent for training/evaluation
 */
export type ScenarioConfig = WallScenario | AgentScenario
export type ScenarioType = 'Wall'
export type ScenarioType1 = 'Agent'
export type WinRound = number
export type LoseRound = number
export type PaddleHit = number
export type NearMissMultiplier = number
export type NearMissExponent = number
export type NearMissMinDistance = number
export type SurvivalRewardMultiplier = number
export type EndurancePenaltyMultiplier = number

export interface TrainingRequest {
  training_id: TrainingId
  training_units: TrainingUnits
  player_config: AgentConfig
  scenario: ScenarioConfig
  reward_config: RewardConfig
  training_output?: TrainingOutput
  [k: string]: unknown
}
/**
 * Describes a configuration of an agent
 */
export interface AgentConfig {
  memory_config: ParentMemoryNodeConfig
  genome_attributes: GenomeAttributes
  [k: string]: unknown
}
export interface RootNode {
  type: 'RootNode'
  [k: string]: unknown
}
export interface TrainedNode {
  type: 'TrainedNode'
  memory: MemoryConfig
  [k: string]: unknown
}
export interface MemoryConfig {
  memory_id: MemoryId
  memory_url: MemoryUrl
  [k: string]: unknown
}
export interface GenomeAttributes {
  size: Size
  strength: Strength
  max_speed: MaxSpeed
  endurance: Endurance
  [k: string]: unknown
}
export interface WallScenario {
  scenario_type: ScenarioType
  [k: string]: unknown
}
export interface AgentScenario {
  scenario_type: ScenarioType1
  agent_config: AgentConfig
  [k: string]: unknown
}
export interface RewardConfig {
  win_round: WinRound
  lose_round: LoseRound
  paddle_hit: PaddleHit
  near_miss_multiplier: NearMissMultiplier
  near_miss_exponent: NearMissExponent
  near_miss_min_distance: NearMissMinDistance
  survival_reward_multiplier: SurvivalRewardMultiplier
  endurance_penalty_multiplier: EndurancePenaltyMultiplier
  [k: string]: unknown
}
/**
 * Output of the previous iteration
 */
export interface TrainingOutput {
  /**
   * The S3 path where the training job saves the model artifacts.
   */
  output_path: string
  /**
   * The number of training units completed.
   */
  completed_training_units: number
  /**
   * Indicates whether the training job has completed, when all training units are  carried out
   */
  is_completed: boolean
  [k: string]: unknown
}